+++ config/freeswitch.php
@@ -303,81 +303,16 @@ return [
         | an explicit granular override is defined in the environment.
         |
         */
-        'cache_ttl' => (int) env(
-            'XML_CACHE_TTL',
-            env('FREESWITCH_XML_HANDLER_CACHE_TTL', 5),
-        ),
-        'dialplan_cache_ttl' => (int) env(
-            'XML_CACHE_DIALPLAN_TTL',
-            env(
-                'FREESWITCH_XML_HANDLER_DIALPLAN_CACHE_TTL',
-                env('XML_CACHE_TTL', env('FREESWITCH_XML_HANDLER_CACHE_TTL', 5)),
-            ),
-        ),
-        'dialplan_contributor_cache_ttl' => (int) env(
-            'XML_CACHE_CONTRIBUTOR_TTL',
-            env(
-                'FREESWITCH_XML_HANDLER_DIALPLAN_CONTRIBUTOR_CACHE_TTL',
-                env('XML_CACHE_TTL', env('FREESWITCH_XML_HANDLER_CACHE_TTL', 5)),
-            ),
-        ),
-        'directory_cache_ttl' => (int) env(
-            'XML_CACHE_DIRECTORY_TTL',
-            env(
-                'FREESWITCH_XML_HANDLER_DIRECTORY_CACHE_TTL',
-                env('XML_CACHE_TTL', env('FREESWITCH_XML_HANDLER_CACHE_TTL', 5)),
-            ),
-        ),
-        'cache_store' => env(
-            'XML_CACHE_STORE',
-            env(
-                'FREESWITCH_XML_HANDLER_CACHE_STORE',
-                env('FREESWITCH_XML_HANDLER_DIALPLAN_CACHE_STORE'),
-            ),
-        ),
-        'dialplan_cache_store' => env(
-            'XML_CACHE_DIALPLAN_STORE',
-            env(
-                'FREESWITCH_XML_HANDLER_DIALPLAN_CACHE_STORE',
-                env(
-                    'XML_CACHE_STORE',
-                    env('FREESWITCH_XML_HANDLER_CACHE_STORE'),
-                ),
-            ),
-        ),
-        'directory_cache_store' => env(
-            'XML_CACHE_DIRECTORY_STORE',
-            env(
-                'FREESWITCH_XML_HANDLER_DIRECTORY_CACHE_STORE',
-                env(
-                    'XML_CACHE_STORE',
-                    env(
-                        'FREESWITCH_XML_HANDLER_CACHE_STORE',
-                        env('FREESWITCH_XML_HANDLER_DIALPLAN_CACHE_STORE'),
-                    ),
-                ),
-            ),
-        ),
-        'acl_cache_ttl' => (int) env(
-            'XML_CACHE_ACL_TTL',
-            env(
-                'FREESWITCH_XML_HANDLER_ACL_CACHE_TTL',
-                env('XML_CACHE_TTL', env('FREESWITCH_XML_HANDLER_CACHE_TTL', 5)),
-            ),
-        ),
-        'acl_cache_store' => env(
-            'XML_CACHE_ACL_STORE',
-            env(
-                'FREESWITCH_XML_HANDLER_ACL_CACHE_STORE',
-                env(
-                    'XML_CACHE_STORE',
-                    env(
-                        'FREESWITCH_XML_HANDLER_CACHE_STORE',
-                        env('FREESWITCH_XML_HANDLER_DIALPLAN_CACHE_STORE'),
-                    ),
-                ),
-            ),
-        ),
+        'cache_ttl' => (int) env('XML_CACHE_TTL', 5),
+        'dialplan_cache_ttl' => (int) env('XML_CACHE_DIALPLAN_TTL', env('XML_CACHE_TTL', 5)),
+        'dialplan_contributor_cache_ttl' => (int) env('XML_CACHE_CONTRIBUTOR_TTL', env('XML_CACHE_TTL', 5)),
+        'directory_cache_ttl' => (int) env('XML_CACHE_DIRECTORY_TTL', env('XML_CACHE_TTL', 5)),
+        'acl_cache_ttl' => (int) env('XML_CACHE_ACL_TTL', env('XML_CACHE_TTL', 5)),
+
+        'cache_store' => env('XML_CACHE_STORE'),
+        'dialplan_cache_store' => env('XML_CACHE_DIALPLAN_STORE', env('XML_CACHE_STORE')),
+        'directory_cache_store' => env('XML_CACHE_DIRECTORY_STORE', env('XML_CACHE_STORE')),
+        'acl_cache_store' => env('XML_CACHE_ACL_STORE', env('XML_CACHE_STORE')),
         'pin_trigger' => env('FREESWITCH_PIN_TRIGGER', '*97'),
         'routing_version_store' => env('FREESWITCH_XML_HANDLER_ROUTING_VERSION_STORE'),